[#3] Refactor: 구매가격, 상품리스트 영역 추상화
